separated survey functionality - cleanup
