+++ dashboard-ui/src/components/Survey/dynamicTemplate.jsx
@@ -146,6 +146,9 @@ export class DynamicTemplate extends SurveyQuestionElementBase {
             userActions: [],
         };
         this.updateActionLogs = this.updateActionLogs.bind(this)
+        console.log(this.question);
+        console.log(this.question.parent);
+        console.log(this.question.parent.data.jsonObj);
         this.scenarioIndex = this.question.parent.data.jsonObj.scenarioIndex
     }
